requirements.txt
dgl-cu101==0.6.1
transformers==4.5.1
torch==1.8.1+cu101
scipy==1.4.1